+++ VoxelGame/src/com/kyry/voxel/world/blocks/BlockStone.java
@@ -7,7 +7,7 @@ import com.nishu.utils.Color4f;
 public class BlockStone extends Block{
 
 	@Override
-	public short getId() {
+	public byte getId() {
 		return 3;
 	}
